# packages/server/openapi.snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "H-Vault API",
    "version": "0.8.0",
    "description": "Zero-knowledge password manager, secret store, and encrypted note-taking API. All vault data is encrypted client-side with AES-256-GCM before reaching the server — the server never sees plaintext user data.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT access token (default 5 min lifetime, configurable via JWT_ACCESS_EXPIRY). Obtain via POST /auth/login."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token",
        "description": "HMAC-SHA256 double-submit CSRF token. Fetch from GET /csrf-token before state-changing requests."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid request body"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["success", "error"]
      },
...
      "BulkReEncryptRequest": {
        "type": "object",
        "required": [
          "authHash",
          "items",
          "newEncryptedVaultKey",
          "newVaultKeyIv",
          "newVaultKeyTag"
        ],
        "properties": {
          "authHash": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Current auth hash for verification"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "encryptedName",
                "nameIv",
                "nameTag",
                "encryptedData",
                "dataIv",
                "dataTag"
              ],
              "properties": {
                "id": {
                  "type": "string"
                },
                "encryptedName": {
                  "type": "string"
                },
                "nameIv": {
                  "type": "string"
                },
                "nameTag": {
                  "type": "string"
                },
                "encryptedData": {
                  "type": "string"
                },
                "dataIv": {
                  "type": "string"
                },
                "dataTag": {
                  "type": "string"
                },
                "searchHash": {
                  "type": "string"
                }
              }
            },
            "maxItems": 10000
          },
          "newEncryptedVaultKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "newVaultKeyIv": {
            "type": "string",
            "minLength": 1,
            "maxLength": 24
          },
          "newVaultKeyTag": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          }
        }
      },
      "FolderResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "encryptedName": {
            "type": "string"
          },
          "nameIv": {
            "type": "string"
          },
          "nameTag": {
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateFolderRequest": {
        "type": "object",
        "required": ["encryptedName", "nameIv", "nameTag"],
        "properties": {
          "encryptedName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "nameIv": {
            "type": "string",
            "minLength": 1,
            "maxLength": 24
          },
          "nameTag": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "parentId": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "maxLength": 20
          },
          "sortOrder": {
            "type": "integer",
            "default": 0
          }
        }
      },
...
      "RestoreBackupRequest": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "conflictStrategy": {
            "type": "string",
            "enum": ["skip", "overwrite", "keep_both"],
            "default": "skip"
          },
          "data": {
            "type": "string",
            "minLength": 1,
            "maxLength": 26214400,
            "description": "Backup file contents (max 25 MB)"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request body failed schema validation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/csrf-token": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get CSRF token",
        "description": "Returns a double-submit CSRF token. Include this token in the `x-csrf-token` header for all state-changing requests (POST, PUT, DELETE). Rate limited by csrfLimiter (30 req/IP per 15 min) in production.",
        "responses": {
          "200": {
            "description": "CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/vault/items/bulk-reencrypt": {
      "post": {
        "tags": ["Vault"],
        "summary": "Bulk re-encrypt vault items",
        "description": "Re-encrypts all vault items with a new vault key after a master password change. Verifies the current auth hash before proceeding. Rate limited: 3 req/IP per 15 min.",
        "security": [
          {
            "bearerAuth": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkReEncryptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Items re-encrypted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "updated": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/folders": {
      "get": {
        "tags": ["Folders"],
        "summary": "List folders",
        "description": "Returns all folders for the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Folder list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FolderResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Folders"],
        "summary": "Create folder",
        "description": "Creates a new folder for organizing vault items.",
        "security": [
          {
            "bearerAuth": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/FolderResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/backup/restore": {
      "post": {
        "tags": ["Backup"],
        "summary": "Restore from backup",
        "description": "Restores vault items and folders from an encrypted backup file. Supports skip, overwrite, and keep_both conflict strategies. Rate limited: 3 req/IP per 15 min.",
        "security": [
          {
            "bearerAuth": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreBackupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Backup restored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "itemsRestored": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    }
  }
}